Extracted utility functions from CompilerApp class
